refactor(llm): extract StructuredOutputMixin (Tier-2.5 Task 3)
Move the response-format selection helpers (_supports_response_schema,
_provider_for_model, _accepts_json_schema_response_format,
_provider_response_format), the post-hoc parse orchestrator
(_maybe_parse_structured_output), and the _JSON_SCHEMA_PROVIDER_ALLOWLIST class
constant into StructuredOutputMixin in _litellm_structured_output.py; compose it
into LiteLLMClient. The mixin self-types config via a per-mixin TYPE_CHECKING
stub. Moves BEFORE text-gen since _provider_response_format and
_maybe_parse_structured_output are the cross-mixin edges text-gen depends on.

SINK-1 (patch-where-used): repoint assert_provider_safe_schema (:1184/:1209) from
the old litellm_client namespace to _litellm_structured_output — the real guard
raises under pytest, so a no-op patch would fail loud. litellm.supports_response_
schema/get_llm_provider stay shared-module attrs.

Bodies verbatim. Facade drops now-unused imports (assert_provider_safe_schema,
strict_response_format_for_model, json, lru_cache) and re-exports the two
test-imported JSON helpers by alias. _make_request / _build_completion_params
reach the moved methods via mixin MRO (self dispatch), unchanged.
